# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/core/province": {
      "post": {
        "description": "เพิ่มข้อมูลจังหวัด",
        "operationId": "ProvinceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProvinceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "ยโสธร"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      },
      "get": {
        "description": "เรียกดูจังหวัดทั้งหมด",
        "operationId": "ProvinceController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": [
                      {
                        "id": 1,
                        "name": "กรุงเทพ"
                      },
                      {
                        "id": 2,
                        "name": "ยโสธร"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      }
    },
    "/api/core/province/search": {
      "get": {
        "description": "เรียกดูจังหวัดตามเงื่อนไข",
        "operationId": "ProvinceController_findByCondition",
        "parameters": [
          {
            "name": "textSearch",
            "required": false,
            "in": "query",
            "description": "Search text (comma-separated for multiple values)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchField",
            "required": false,
            "in": "query",
            "description": "Fields to search in (comma-separated for multiple fields)",
            "schema": {
              "example": "id,name",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": "10",
              "type": "string"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": [
                      {
                        "id": 1,
                        "name": "กรุงเทพ"
                      },
                      {
                        "id": 2,
                        "name": "ยโสธร"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      }
    },
    "/api/core/province/{id}": {
      "get": {
        "description": "เรียกดูจังหวัดตาม id",
        "operationId": "ProvinceController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "name": "กรุงเทพ"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      },
      "patch": {
        "description": "แก้ไขข้อมูลจังหวัด",
        "operationId": "ProvinceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProvinceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "ยโสธร2"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      },
      "delete": {
        "description": "ลบจังหวัดตาม id",
        "operationId": "ProvinceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "กรุงเทพ"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      }
    },
    "/api/core/province/delete-many": {
      "post": {
        "description": "ลบจังหวัดหลายแถวด้วย id แบบ array",
        "operationId": "ProvinceController_removeMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteManyProvinceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "count": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Province"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Title",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://nakdev.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-version": {
        "type": "apiKey",
        "in": "header",
        "name": "x-version"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProvinceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "ชื่อจังหวัด"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "UpdateProvinceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "ชื่อจังหวัด"
          },
          "id": {
            "type": "number",
            "description": "รหัสจังหวัด",
            "example": 2
          }
        },
        "required": [
          "id"
        ]
      },
      "DeleteManyProvinceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "array",
            "description": "รหัสจังหวัด",
            "example": [
              1,
              2,
              5,
              8
            ]
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}